pyproject.toml
[project]
name = "ezdoc-back"
version = "0.1.0"
description = ""
authors = [
    {name = "Renzo Telenta",email = "renzotelenta@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.122.0,<0.123.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "psycopg2-binary (>=2.9.0,<3.0.0)",
    "docxtpl (>=0.16.7,<1.0.0)",
    "jinja2 (>=3.1.0,<4.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "requests (>=2.31.0,<3.0.0)"
]

[tool.poetry]
packages = [{include = "app"}]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
